docs: add mention of the React library under E2 introduction
Since the React library has been migrated under the Paytrail account as decided, we need to have it briefly featured in the documentation as well. Reordered sections to describe packages per language, but in the future it might be better to have own page for all the open-source libraries. For now, our focus is on E2 interface so let's keep them here.